Introducing save to legacy functionality
Add low level function 'save_legacy', which is called by high level version.
The low level gets data in a abstract matrix and the name of the variables.
The high level uses SpatialData variable, forming the data matrix and variable names
from the domain (PointSet or RegularGrid).
A variant with variable names as String is also provided.